Rename to InitBrigade and minor changes to defaults
