+++ brigade/plugins/inventory/simple.py
@@ -102,7 +102,7 @@ class SimpleInventory(Inventory):
         cmh:
             group: all
     """
-    def __init__(self, host_file="hosts.yaml", group_file=None, **kwargs):
+    def __init__(self, host_file="hosts.yaml", group_file="groups.yaml", **kwargs):
         with open(host_file, "r") as f:
             hosts = yaml.load(f.read())
