+++ tests/core/test_InitBrigade/a_config.yaml
